refactor(web): split the Transactions and Portfolio pages
Both were ~1280-line components where the page, its view-model derivation
and its forms were interleaved. Neither behaves differently.

Transactions.tsx  1269 -> 749
  transactionsHelpers.ts    constants, shapes, pure helpers
  useTransactionHistory.ts  trades + transactions reconciled, filtered,
                            grouped and charted
  LogTransactionDialog.tsx  the log/transfer modal

Portfolio.tsx     1293 -> 1057
  portfolioHelpers.ts       shapes, colours, the money formatter
  usePortfolioData.ts       per-tab stats, equity curve, dividends,
                            sorted holdings

Portfolio is still over 1000: its remaining bulk is five inline form
panels in the JSX, which are a separate extraction and not attempted here.

Two things fell out of the move rather than being aimed at:

The Portfolio page had an early `if (!activeHousehold) return <div/>`
sitting *between* two useMemo calls, so a render without a household ran
a different number of hooks than one with it. With the memos inside the
hook the guard now follows them, which is where it always belonged.

`formatCurrency` was a component-local closure that the extracted
derivation also needed. It's a shared factory now, so the page and the
view model can't drift on decimal places.

Transactions had no test at all, which is a poor thing to discover while
moving 240 lines of its JSX. Added a smoke test that renders the page and
opens the dialog on both tabs; it fails if the tab callback loses its
binding, which is exactly the mistake the extraction could have made.
