+++ frontend/src/pages/Portfolio/Portfolio.tsx
@@ -1,4 +1,4 @@
-import { useState, useEffect, useMemo } from "react"
+import { useState, useEffect } from "react"
 import { useLoaderData, useRevalidator, useNavigation, useSearchParams, Link } from "react-router"
 import { StatCard } from "../../components/ui/StatCard"
 import { Card, CardContent, CardHeader, CardTitle, CardDescription } from "../../components/ui/Card"
@@ -6,7 +6,7 @@ import { Badge } from "../../components/ui/Badge"
 import { Button } from "../../components/ui/Button"
 import { Input } from "../../components/ui/Input"
 import { Select } from "../../components/ui/Select"
-import { cn, returnBasis } from "../../lib/utils"
+import { cn } from "../../lib/utils"
 import { Area, AreaChart, ResponsiveContainer, Tooltip, XAxis, YAxis, CartesianGrid } from "recharts"
 import { useHousehold } from "../../lib/HouseholdContext"
 import api from "../../lib/api"
@@ -14,43 +14,11 @@ import type { SubPortfolioResponse, AssetResponse } from "../../types/types"
 import type { PortfolioLoaderData } from "./portfolio.loader"
 import { TimeframeSelector } from "../../components/ui/TimeframeSelector"
 import { TopBar } from "../../components/TopBar"
+import { ALLOCATION_COLORS, makeCurrencyFormatter, type Holding } from "./portfolioHelpers"
+import { usePortfolioData } from "./usePortfolioData"
 
 export { portfolioLoader as loader } from "./portfolio.loader";
 
-type Holding = {
-    assetId: string;
-    ticker: string;
-    name: string;
-    shares: number;
-    avgCost: number; // Home currency
-    currentPrice: number; // Home currency
-    currency: string; // Ticker's base currency
-    avgCostNative: number;
-    currentPriceNative: number;
-    assetType: string;
-    pricingMode: "market" | "manual";
-};
-
-const ALLOCATION_COLORS = ["#38bdf8", "#4ade80", "#fbbf24", "#e879f9", "#f472b6", "#a78bfa", "#fb923c", "#2dd4bf"];
-
-type PortfolioData = {
-    stats: {
-        equity: string;
-        unrealized: string;
-        unrealizedPercent: number;
-        realized: string;
-        sharpe: string;
-        sortino: string;
-        treynor: string;
-        alpha: string;
-        beta: string;
-        drawdown: string;
-        twr: string;
-        irr: string;
-    };
-    history: any[];
-    holdings: Holding[];
-};
 
 
 
@@ -62,12 +30,7 @@ export default function Portfolio() {
     const [searchParams] = useSearchParams()
     const startDate = searchParams.get("start_date")
 
-    const formatCurrency = (val: number, code?: string) => new Intl.NumberFormat('en-US', {
-        style: 'currency',
-        currency: code || activeHousehold?.base_currency || 'USD',
-        minimumFractionDigits: 2,
-        maximumFractionDigits: 2
-    }).format(val);
+    const formatCurrency = makeCurrencyFormatter(activeHousehold?.base_currency);
 
     const formatCompactCurrency = (val: number) => new Intl.NumberFormat('en-US', {
         style: 'currency',
@@ -286,184 +249,16 @@ export default function Portfolio() {
         }
     };
 
-    const portfoliosData = useMemo(() => {
-        const dataMap: Record<string, PortfolioData> = {};
+    // Snapshots, trades and metrics reconciled per tab into what the page renders.
+    const { portfoliosData, dividendData, sortedHoldings, rawData } = usePortfolioData({
+        latestSnapshots, timeseries, assets, subportfolios, metrics, dividends,
+        activeTab, timeframe, sortConfig,
+        baseCurrency: activeHousehold?.base_currency, startDate,
+    })
 
-        const binHistory = (history: any[]) => {
-            if (timeframe === "Daily") return history;
-
-            const binned = new Map<string, number>();
-            history.forEach(item => {
-                const d = new Date(item.date);
-                let key = "";
-                if (timeframe === "Weekly") {
-                    const startOfWeek = new Date(d);
-                    const day = d.getDay();
-                    const diff = d.getDate() - day + (day === 0 ? -6 : 1); // Monday
-                    startOfWeek.setDate(diff);
-                    key = startOfWeek.toISOString().split('T')[0];
-                } else if (timeframe === "Monthly") {
-                    key = `${d.getFullYear()}-${String(d.getMonth() + 1).padStart(2, '0')}-01`;
-                } else if (timeframe === "Yearly") {
-                    key = `${d.getFullYear()}-01-01`;
-                }
-                binned.set(key, item.equity);
-            });
-
-            return Array.from(binned.entries())
-                .filter(([date]) => !startDate || date >= startDate)
-                // ⚡ Bolt: Fast string comparison instead of localeCompare
-                .sort((a, b) => (a[0] < b[0] ? -1 : (a[0] > b[0] ? 1 : 0)))
-                .map(([date, equity]) => ({ date, equity }));
-        };
-
-        const computeStats = (hlds: Holding[], realizedPnL: number, history: any[], equity: number, m?: any): PortfolioData => {
-            let costBasis = 0;
-            for (const h of hlds) {
-                costBasis += h.shares * h.avgCost;
-            }
-            const unrealized = equity - costBasis;
-            const unrealizedPercent = costBasis > 0 ? (unrealized / costBasis) * 100 : 0;
-
-            const formatPercent = (val: number) => `${(val * 100).toFixed(2)}%`;
-            const sign = (val: number) => val >= 0 ? '+' : '';
-
-            if (history.length === 0) {
-                history = [{ date: new Date().toLocaleDateString('default', { month: 'short', day: 'numeric', year: 'numeric' }), equity: 0 }];
-            }
-
-            return {
-                stats: {
-                    equity: formatCurrency(equity),
-                    unrealized: `${sign(unrealized)}${formatCurrency(Math.abs(unrealized))}`,
-                    unrealizedPercent: (m?.simple_return ?? 0) * 100 || unrealizedPercent,
-                    realized: `${sign(realizedPnL)}${formatCurrency(Math.abs(realizedPnL))}`,
-                    sharpe: m?.sharpe_ratio !== undefined && m?.sharpe_ratio !== null ? m.sharpe_ratio.toFixed(2) : "N/A",
-                    sortino: m?.sortino_ratio !== undefined && m?.sortino_ratio !== null ? m.sortino_ratio.toFixed(2) : "N/A",
-                    treynor: m?.treynor_ratio !== undefined && m?.treynor_ratio !== null ? m.treynor_ratio.toFixed(2) : "N/A",
-                    alpha: m?.alpha !== undefined && m?.alpha !== null ? formatPercent(m.alpha) : "N/A",
-                    beta: m?.beta !== undefined && m?.beta !== null ? m.beta.toFixed(2) : "N/A",
-                    drawdown: m?.volatility !== undefined && m?.volatility !== null ? formatPercent(m.volatility) : "N/A",
-                    twr: m?.time_weighted_return !== undefined && m?.time_weighted_return !== null ? formatPercent(m.time_weighted_return) : "N/A",
-                    irr: m?.money_weighted_return !== undefined && m?.money_weighted_return !== null ? formatPercent(m.money_weighted_return) : "N/A",
-                    returnBasis: returnBasis(m?.annualized)
-                },
-                history,
-                holdings: hlds
-            };
-        };
-
-        // Equity curve from pre-aggregated (date, sub_portfolio) totals — already summed
-        // across assets server-side, so this just sorts/bins/windows.
-        const buildHistory = (points: typeof timeseries) => {
-            const history = points
-                .slice()
-                .sort((a, b) => (a.date < b.date ? -1 : (a.date > b.date ? 1 : 0)))
-                .map(t => ({ date: t.date, equity: Number(t.total_value_home_currency) }))
-                .filter(item => !startDate || item.date >= startDate);
-            return binHistory(history);
-        };
-
-        // Holdings table from the latest-date-only per-asset rows (server-filtered via
-        // latest_only=true, so every row here is already "current").
-        const buildHoldings = (latestSnaps: typeof latestSnapshots) => {
-            const assetMap = new Map(assets.map(a => [a.id, a]));
-            const holdingMap = new Map<string, Holding>();
-
-            latestSnaps
-                .filter(s => s.quantity > 0.001)
-                .forEach(s => {
-                    const asset = assetMap.get(s.asset_id);
-                    const ticker = asset?.ticker || "UNKNOWN";
-                    const currency = asset?.currency || "USD";
-                    
-                    const currentPriceHome = Number(s.price) * (s.exchange_rate_used || 1.0);
-                    const costBasisHome = Number(s.average_cost_basis_home_currency ?? (Number(s.average_cost_basis) * (s.exchange_rate_used || 1.0)));
-                    const currentPriceNative = Number(s.price);
-                    const costBasisNative = Number(s.average_cost_basis);
-
-                    if (holdingMap.has(ticker)) {
-                        const existing = holdingMap.get(ticker)!;
-                        const totalShares = existing.shares + s.quantity;
-                        
-                        // Weighted average for home currency
-                        const totalCostHome = (existing.shares * existing.avgCost) + (s.quantity * costBasisHome);
-                        existing.avgCost = totalShares > 0 ? totalCostHome / totalShares : 0;
-                        
-                        // Weighted average for native currency
-                        const totalCostNative = (existing.shares * existing.avgCostNative) + (s.quantity * costBasisNative);
-                        existing.avgCostNative = totalShares > 0 ? totalCostNative / totalShares : 0;
-                        
-                        existing.shares = totalShares;
-                        existing.currentPrice = currentPriceHome;
-                        existing.currentPriceNative = currentPriceNative;
-                    } else {
-                        holdingMap.set(ticker, {
-                            assetId: s.asset_id,
-                            ticker: ticker,
-                            name: asset?.name || "Unknown Asset",
-                            shares: s.quantity,
-                            avgCost: costBasisHome,
-                            currentPrice: currentPriceHome,
-                            currency: currency,
-                            avgCostNative: costBasisNative,
-                            currentPriceNative: currentPriceNative,
-                            assetType: asset?.type || "other",
-                            pricingMode: asset?.pricing_mode || "market"
-                        });
-                    }
-                });
-
-            const holdings: Holding[] = Array.from(holdingMap.values());
-
-            const currentEquity = latestSnaps.reduce((sum, s) => sum + Number(s.current_value_home_currency), 0);
-
-            return { holdings, currentEquity };
-        };
-
-        // Overall
-        const overallHoldings = buildHoldings(latestSnapshots);
-        dataMap["Overall"] = computeStats(
-            overallHoldings.holdings,
-            0,
-            buildHistory(timeseries),
-            overallHoldings.currentEquity,
-            metrics?.overall_metrics
-        );
-
-        // Subportfolios
-        for (const sp of subportfolios) {
-            const spHoldings = buildHoldings(latestSnapshots.filter(s => s.sub_portfolio_id === sp.id));
-            const spMetric = metrics?.sub_portfolio_metrics.find(m => m.sub_portfolio_id === sp.id);
-            dataMap[sp.name] = computeStats(
-                spHoldings.holdings,
-                0,
-                buildHistory(timeseries.filter(t => t.sub_portfolio_id === sp.id)),
-                spHoldings.currentEquity,
-                spMetric?.metrics
-            );
-        }
-
-        return dataMap;
-    }, [latestSnapshots, timeseries, assets, subportfolios, timeframe, metrics, startDate]);
-
-    // Dividends for the active tab (all sub-portfolios when "Overall"), newest first.
-    // Kept above the early return so hook order stays stable across renders.
-    const dividendData = useMemo(() => {
-        const sp = subportfolios.find(s => s.name === activeTab);
-        const assetById = new Map(assets.map(a => [a.id, a]));
-        const relevant = dividends
-            .filter(d => activeTab === "Overall" || d.sub_portfolio_id === sp?.id)
-            .map(d => ({
-                ...d,
-                ticker: assetById.get(d.asset_id)?.ticker || "UNKNOWN",
-                homeAmount: d.amount_home_currency ?? d.amount,
-            }))
-            .sort((a, b) => (a.date < b.date ? 1 : (a.date > b.date ? -1 : 0)));
-        const total = relevant.reduce((sum, d) => sum + (Number(d.homeAmount) || 0), 0);
-        return { rows: relevant, total };
-    }, [dividends, assets, subportfolios, activeTab]);
-
+    // After the hook, never before it: this guard used to sit between two
+    // `useMemo` calls, so a render without a household ran a different number of
+    // hooks than one with it.
     if (!activeHousehold) {
         return (
             <div className="flex-1 flex items-center justify-center p-8 text-base-500">
@@ -474,37 +269,6 @@ export default function Portfolio() {
 
     const isLoading = navigation.state === "loading" || revalidator.state === "loading";
 
-    // Default to Overall if tab is deleted
-    const rawData = portfoliosData[activeTab] || portfoliosData["Overall"];
-    
-    const sortedHoldings = useMemo(() => {
-        if (!rawData.holdings) return [];
-        const sortable = [...rawData.holdings];
-        if (sortConfig !== null) {
-            sortable.sort((a, b) => {
-                let aValue: any;
-                let bValue: any;
-
-                switch (sortConfig.key) {
-                    case 'asset': aValue = a.ticker; bValue = b.ticker; break;
-                    case 'shares': aValue = a.shares; bValue = b.shares; break;
-                    case 'avgCost': aValue = a.avgCost; bValue = b.avgCost; break;
-                    case 'price': aValue = a.currentPrice; bValue = b.currentPrice; break;
-                    case 'value': aValue = a.shares * a.currentPrice; bValue = b.shares * b.currentPrice; break;
-                    case 'return': 
-                        aValue = (a.shares * a.currentPrice) - (a.shares * a.avgCost);
-                        bValue = (b.shares * b.currentPrice) - (b.shares * b.avgCost);
-                        break;
-                    default: aValue = 0; bValue = 0;
-                }
-
-                if (aValue < bValue) return sortConfig.direction === 'asc' ? -1 : 1;
-                if (aValue > bValue) return sortConfig.direction === 'asc' ? 1 : -1;
-                return 0;
-            });
-        }
-        return sortable;
-    }, [rawData.holdings, sortConfig]);
 
     const requestSort = (key: string) => {
         let direction: 'asc' | 'desc' = 'asc';
